Minimal Working FrontController + CLI platform
- Re-implements bindings for the ev5 platform to align with the EV6
names.
- Sends a CAN FC status message (currently always 0) to tell LV that the
system is on.
- Removes FreeRTOS because we don't want to worry about that for now.
We're only running a single task anyways. Reconsider in the future as FC
becomes more complicated.
- This means we could implement the CLI platform which helped verify
that our state machine was working.

+++ firmware/projects/FrontController/inc/app.hpp
@@ -12,10 +12,10 @@ class AnalogInput {
 
 public:
     AnalogInput(shared::periph::AnalogInput& adc,
-                shared::util::Mapper<double, uint16_t>* adc_to_position)
+                shared::util::Mapper<float, uint16_t>* adc_to_position)
         : adc_(adc), adc_to_position_(adc_to_position) {}
 
-    double Update() {
+    float Update() {
         uint32_t position = adc_.ReadVoltage();
         moving_average_.LoadValue(uint16_t(position));
         return GetPosition();
@@ -24,12 +24,12 @@ public:
     /**
      * @brief Get the position from the last `Update()` call.
      */
-    inline double GetPosition() {
+    inline float GetPosition() {
         return adc_to_position_->Evaluate(moving_average_.GetValue());
     }
 
 private:
     shared::periph::AnalogInput& adc_;
     shared::util::MovingAverage<uint16_t, kMovingAverageLength> moving_average_;
-    const shared::util::Mapper<double, uint16_t>* adc_to_position_;
+    const shared::util::Mapper<float, uint16_t>* adc_to_position_;
 };
